requirements.txt
streamlit==1.32.0
pandas==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.12.0
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
pytest==7.4.0